+++ library/Collections/PhysInterp1DScaled.cpp
@@ -62,8 +62,8 @@ using LibUtilities::eTriangle;
 class PhysInterp1DScaled_Helper : virtual public Operator
 {
 public:
-    void UpdateFactors(StdRegions::FactorMap factors,
-                       [[maybe_unused]] int coll_phys_offset) override
+    void UpdateFactors(StdRegions::FactorMap factors) override
+
     {
         if (factors == m_factors)
         {
@@ -142,8 +142,7 @@ public:
         ASSERTL0(false, "Not valid for this operator.");
     }
 
-    void UpdateFactors(StdRegions::FactorMap factors,
-                       [[maybe_unused]] int coll_phys_offset) override
+    void UpdateFactors(StdRegions::FactorMap factors) override
     {
         if (factors == m_factors)
         {
@@ -223,8 +222,7 @@ public:
                  "BwdTrans_MatrixFree: Not valid for this operator.");
     }
 
-    void UpdateFactors([[maybe_unused]] StdRegions::FactorMap factors,
-                       [[maybe_unused]] int coll_phys_offset) override
+    void UpdateFactors([[maybe_unused]] StdRegions::FactorMap factors) override
     {
         ASSERTL0(false, "Not valid for this operator.");
     }
@@ -302,8 +300,7 @@ public:
         ASSERTL0(false, "Not valid for this operator.");
     }
 
-    void UpdateFactors([[maybe_unused]] StdRegions::FactorMap factors,
-                       [[maybe_unused]] int coll_phys_offset) override
+    void UpdateFactors([[maybe_unused]] StdRegions::FactorMap factors) override
     {
         m_factors = factors;
     }
@@ -493,10 +490,9 @@ public:
         ASSERTL0(false, "Not valid for this operator.");
     }
 
-    void UpdateFactors([[maybe_unused]] StdRegions::FactorMap factors,
-                       [[maybe_unused]] int coll_phys_offset) override
+    void UpdateFactors([[maybe_unused]] StdRegions::FactorMap factors) override
     {
-        PhysInterp1DScaled_Helper::UpdateFactors(factors, coll_phys_offset);
+        PhysInterp1DScaled_Helper::UpdateFactors(factors);
         m_factors = factors;
     }